etl/requirements.txt
cassandra-driver==3.28.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
geomet==0.2.1.post1
jwt==1.3.1
pycparser==2.21
PyJWT==2.8.0
six==1.16.0
